Rename HnswIndexBase -> HnswIndex and remove the templated class.
