+++ searchlib/src/tests/tensor/hnsw_index/hnsw_index_test.cpp
@@ -43,15 +43,14 @@ struct LevelGenerator : public RandomLevelGenerator {
 
 using FloatVectors = MyDocVectorAccess<float>;
 using FloatSqEuclideanDistance = SquaredEuclideanDistance<float>;
-using FloatIndex = HnswIndex<float>;
-using FloatIndexUP = std::unique_ptr<FloatIndex>;
+using HnswIndexUP = std::unique_ptr<HnswIndex>;
 
 class HnswIndexTest : public ::testing::Test {
 public:
     FloatVectors vectors;
     FloatSqEuclideanDistance distance_func;
     LevelGenerator level_generator;
-    FloatIndexUP index;
+    HnswIndexUP index;
 
     HnswIndexTest()
         : vectors(),
@@ -63,8 +62,8 @@ public:
                .set(7, {3, 5});
     }
     void init(bool heuristic_select_neighbors) {
-        index = std::make_unique<FloatIndex>(vectors, distance_func, level_generator,
-                                             HnswIndexBase::Config(2, 1, 10, heuristic_select_neighbors));
+        index = std::make_unique<HnswIndex>(vectors, distance_func, level_generator,
+                                            HnswIndex::Config(2, 1, 10, heuristic_select_neighbors));
     }
     void add_document(uint32_t docid, uint32_t max_level = 0) {
         level_generator.level = max_level;
