+++ searchlib/src/vespa/searchlib/tensor/CMakeLists.txt
@@ -6,7 +6,6 @@ vespa_add_library(searchlib_tensor OBJECT
     dense_tensor_store.cpp
     generic_tensor_attribute.cpp
     generic_tensor_store.cpp
-    hnsw_index_base.cpp
     hnsw_index.cpp
     imported_tensor_attribute_vector.cpp
     imported_tensor_attribute_vector_read_guard.cpp
